Fix bad type introduced in one of the previous commits concerning cppcheck issues

+++ modules/tracker/tt_mi/src/mi/vpTemplateTrackerMIInverseCompositional.cpp
@@ -267,7 +267,7 @@ void vpTemplateTrackerMIInverseCompositional::trackNoPyr(const vpImage<unsigned 
 
   do
   {
-    double Nbpoint=0;
+    int Nbpoint=0;
     MIprec=MI;
     MI=0;
